Correcting face button mapping. Fully tested.
In spite of now having the option of changing the button layout in the menu this is still necessary. The easiest way to see the need is running an SNES emulator and realizing that X and A are still Y and B. Please approve and let me know if you have any questions. I'm also testing a few other tweaks but this seems immediately needed.